% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{DSC_proclus}
\alias{DSC_proclus}
\title{ProClus algorithm for use with DSC_ThreeStage}
\usage{
DSC_proclus(numOfClusters = 5, avgDimensions = 3)
}
\arguments{
\item{numOfClusters}{Number of Clusters to be found.}

\item{avgDimensions}{Average number of dimensions in which each cluster
 resides
dsc <- DSC_ThreeStage(macro=DSC_proclus(),micro=DSC_subspaceCluStream())
dsd <- DSD_RandomRBFSubspaceGeneratorEvents()
update(dsc,dsd,1000)}
}
\description{
An implementation of the ProClus algorithm 
that can be used with \link{DSC_ThreeStage}.
For more details on this
algorithm, consult \link[subspace]{ProClus}.
}

