% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{DSC_subclu}
\alias{DSC_subclu}
\title{SubClu algorithm for use with DSC_ThreeStage}
\usage{
DSC_subclu(epsilon = 0.05, minSupport = 50, minOutputDim = 3)
}
\arguments{
\item{epsilon}{this parameter determines the size of the epsilon environment for the DBSCAN that is run as a part of this algorithm.}

\item{minSupport}{minimum number of points in the epsilon environment.}

\item{minOutputDim}{minimum dimensionality that a cluster must have to be output.
dsc <- DSC_ThreeStage(macro=DSC_subclu(),micro=DSC_subspaceCluStream())
dsd <- DSD_RandomRBFSubspaceGeneratorEvents()
update(dsc,dsd,1000)}
}
\description{
An implementation of the SubClu algorithm 
that can be used with \link{DSC_ThreeStage}.
For more details on this
algorithm, consult \link[subspace]{SubClu}
}

