% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\name{DSC_subspaceCluStream}
\alias{DSC_subspaceCluStream}
\title{CluStream for use with DSC_ThreeStage}
\usage{
DSC_subspaceCluStream(timeWindow = 1000, maxNumKernels = 200,
  kernelRadiFactor = 2, streamSpeed = 1)
}
\arguments{
\item{timeWindow}{window for aging the points}

\item{maxNumKernels}{maximum number of microclusters to be produced}

\item{kernelRadiFactor}{multiplier for  the kernel radius}

\item{streamSpeed}{number of points processed per time unit}
}
\description{
A version of the \link[streamMOA]{DSC_CluStream} algorithm that is optimized for use with \link{DSC_ThreeStage}. 
Do not attempt to use this as a standalone stream clustering algorithm.
}
