% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\name{DSC_subspaceDenStream}
\alias{DSC_subspaceDenStream}
\title{DenStream for use with DSC_ThreeStage}
\usage{
DSC_subspaceDenStream(horizon = 1000, epsilon = 0.04, minPoints = 4,
  beta = 0.2, mu = 1, initPoints = 1000, speed = 100)
}
\arguments{
\item{horizon}{range of the window}

\item{epsilon}{defines the epsilon neighborhood}

\item{minPoints}{minimal number of points a cluster has to contain}

\item{beta}{multiplier for mu to detect outlier micro-clusters}

\item{mu}{minimal number of points to form a micro-cluster}

\item{initPoints}{number of points used to initialize the algorithm}

\item{speed}{number of data points processed in one time unit}
}
\description{
A version of the \link[streamMOA]{DSC_DenStream} algorithm that is optimized for use with \link{DSC_ThreeStage}. 
Do not attempt to use this as a standalone stream clustering algorithm.
}
