% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_SubspaceARFFStream.R
\name{DSD_SubspaceARFFStream}
\alias{DSD_SubspaceARFFStream}
\title{Stream Subspace Instances from Disc}
\usage{
DSD_SubspaceARFFStream(file)
}
\arguments{
\item{file}{a string that contains the path to the file}
}
\description{
A DSD object to stream Data Points from a .arff file.
}

