% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_subspace}
\alias{evaluate_subspace}
\title{Evaluate Subspace Clusterings}
\usage{
evaluate_subspace(dsc, dsd, n = 1000, measures = c("clustering error",
  "cmm subspace", "entropy subspace", "f1 subspace", "purity",
  "rand statistic"), alsoTrainOn = F)
}
\arguments{
\item{dsc}{The clusterer whose current clustering should be evaluated.}

\item{dsd}{The stream from which the data points for evaluation should be drawn.}

\item{n}{How many points to evaluate over}

\item{measures}{A vector of evaluation measures to use. By default, all supported measures are used.}

\item{alsoTrainOn}{This will train the clusterer on the data points before running the evaluation.}
}
\description{
This function evaluates Subspace Clusterings based on data points from a stream.
}
