% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stream_interactive}
\alias{plot_stream_interactive}
\title{Show Stream Clustering.}
\usage{
plot_stream_interactive(dsc, points,
  launch.browser = getOption("shiny.launch.browser", interactive()))
}
\arguments{
\item{dsc}{a DSC object representing the clustering of a data stream.}

\item{points}{a \link{data.frame} of points that will be plotted along with
the clustering.}

\item{launch.browser}{will be passed on to \link[shiny]{runApp}, so that the
visualisation can be shown in e.g. RStudio's Viewer pane, if this is
desired.}
}
\description{
A non-animated version of \link{animate_stream_interactive}.
}
